Extract varisat-checker and varisat-lrat crates
