+++ varisat-checker/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "varisat-dimacs"
+name = "varisat-checker"
 version = "0.2.0"
 authors = ["Jannis Harder <me@jix.one>"]
 edition = "2018"
-description = "DIMCAS CNF parser and writer for the Varisat SAT solver"
+description = "Proof checker for proofs generate by the Varisat SAT solver"
 homepage = "https://jix.one/project/varisat/"
 repository = "https://github.com/jix/varisat"
 license = "MIT/Apache-2.0"
@@ -11,18 +11,21 @@ readme = "README.md"
 
 [dependencies]
 failure = "0.1.5"
-itoa = "0.4.4"
+hashbrown = "0.3.0"
+smallvec = "0.6.9"
+log = "0.4.6"
+
     [dependencies.varisat-formula]
     path = "../varisat-formula"
     version = "=0.2.0"
 
+    [dependencies.varisat-dimacs]
+    path = "../varisat-dimacs"
+    version = "=0.2.0"
+
+    [dependencies.varisat-internal-proof]
+    path = "../varisat-internal-proof"
+    version = "=0.2.0"
+
 [dev-dependencies]
 proptest = "0.9.3"
-rand = "0.6.5"
-tempfile = "3.0.7"
-env_logger = "0.6.1"
-
-    [dev-dependencies.varisat-formula]
-    path = "../varisat-formula"
-    version = "=0.2.0"
-    features = ["proptest-strategies", "internal-testing"]
