+++ varisat-cli/Cargo.toml
@@ -12,13 +12,19 @@ edition = "2018"
 autobins = false
 
 [dependencies]
-varisat = { path = "../varisat", version = "0.2.0" }
-
 clap = { version = "2.33.0", features = ["wrap_help"] }
 env_logger = "0.6.1"
 failure = "0.1.5"
 log = "0.4.6"
 toml = "0.5.1"
+
+    [dependencies.varisat]
+    path = "../varisat"
+    version = "=0.2.0"
+
+    [dependencies.varisat-lrat]
+    path = "../varisat-lrat"
+    version = "=0.2.0"
 
 [[bin]]
 name = "varisat"
