+++ varisat-formula/Cargo.toml
@@ -13,10 +13,12 @@ readme = "README.md"
 default = []
 
 proptest-strategies = ["proptest"]
-internal-test-macros = []
+internal-testing = ["proptest", "rand"]
 
 [dependencies]
 proptest = { version = "0.9.3", optional = true }
+rand = { version = "0.6.5", optional = true }
 
 [dev-dependencies]
 proptest = "0.9.3"
+rand = "0.6.5"
