+++ varisat/src/lib.rs
@@ -8,8 +8,6 @@
 //! [cnf]: https://en.wikipedia.org/wiki/Conjunctive_normal_form
 //! [user manual]: https://jix.github.io/varisat/manual/0.2.0/
 
-#[macro_use]
-pub mod checker;
 pub mod config;
 pub mod solver;
 
@@ -29,12 +27,15 @@ mod simplify;
 mod state;
 mod tmp;
 
-#[cfg(test)]
-mod test;
-
+pub use solver::{ProofFormat, Solver};
 pub use varisat_formula::{cnf, lit, CnfFormula, Lit, Var};
 
 pub mod dimacs {
     //! DIMCAS CNF parser and writer.
     pub use varisat_dimacs::*;
+}
+
+pub mod checker {
+    //! Proof checker for Varisat proofs.
+    pub use varisat_checker::{CheckedProofStep, Checker, CheckerError, ProofProcessor};
 }
